Update
Modified database shop table to include the county and city where the shop is located.
Modified UI for creating a new shop
Modified the function and php script for creating a new shop to include the county and city.
Now when creating a new shop a folder is created as well for the images of the shop to be imported to.
The function responsible for showing the images of the shop now selects them from the shop's folder.
The function responsible for uploading the images of the shop was modified as a result of all these modifications.