UPDATE
Created an appointment system for the users to create program the moment when they are going to get a new cut.
The appointments are saved to database and taken from it.
The hours and minutes shown to the users are taken from the shop's working program.
If there is already an appointment for a specific moment then the button becomes red and not interactable to show the user that it is not available.
Barbers can update their pace of finishing a cut.(how much time they need/take to finish a cut)
Minutes shown to the users are created based on the pace of the barber selected so for example if the barber has a 10 minute pace there will be shown minutes from 0-50 increasing from 0 by 10.