UPDATE
Added the option for the barbers to see the calendar of the current month and if there is are appointments in one of those days it the day becomes red so the barber knows there is an appointment in that day.
Barbers can then select the day and all the appointments will show up.
He has the option to select the appointments to check the info left by the client and delete them.