Update
Uploaded the database and php scripts to a online server.
Added the option for the saloon boss to select if he wants his barbers(employees) to have their own prices and let them edit as they wish or if he wants to have a general price for the saloon that will be shown for every employee.
Added the option for the boss to edit the price.
Added the option for the employees to edit the price.
The barber can now see the price of an appointment when looking for appoinments in his barber menu.