UPDATE
Added notifications for the user before the appointment (as a reminder) and after the appointment  (to review his barber if he wishes so).
If the user opens the app through the review notification a menu will be shown where he can choose the barber's rating which will be updated and shown in the Shop's barber list.

Updated the translation.
Made the app look more clean.
Updated the menu bar.