Update
Shop owners can now create custom services for their shops. They can name them in romanian or english for now ( or just one of them ) and when the user wants to create an appointment those services show with the name depending on the language of the app. If the app is in romanian and the service has no name for romanian it will show the english one.